basic server implementation

+++ backend/src/app.ts
@@ -2,6 +2,7 @@ import chalk from "chalk"
 import { debugPrint, figletText, iniPrint, print, printDevOptions } from "./utility/print.js"
 import config, { DEV_OPTIONS } from "./utility/config.js"
 import { createDatabase, deleteDatabase, doesDBExist } from "./database/initialise.js"
+import { Server } from "./server/server.js"
 
 const printWelcome = () => {
   print(figletText(config.NAME) +" v"+ chalk.underline(config.VERSION))
@@ -13,7 +14,7 @@ const printWelcome = () => {
 }
 
 const initaliseDatabase = () => {
-  if(!DEV_OPTIONS.PERSIST_DB) {
+  if(!DEV_OPTIONS.PERSIST_DB && config.DEV_MODE) {
     debugPrint("Removing previous database")
     deleteDatabase()
   }
@@ -21,12 +22,12 @@ const initaliseDatabase = () => {
     iniPrint("Initalising database")
     createDatabase()
   }else iniPrint("Database already exists... skipping")
-  iniPrint("Initalisation Complete\n")
 }
 
 const run = () => {
   printWelcome()
   initaliseDatabase()
+  Server.getInstance().listen()
 }
 
 run()
